Refined AddScriptsToPrefab & AddTightBoxColliderToMeshObject for the Parent Model
Modified the ModelImporter's AddScriptsToPrefab method to selectively add the ObjectManipulator and NearInteractionGrabbable scripts solely to the parent model.

Additionally, updated the AddTightBoxColliderToMeshObject method to double the parent model's box collider height along the y-axis. This adjustment facilitates easier grabbing from above and aids in aligning the model with the patient.