Fix: Correctly update spine material color when toggling visibility off/on
The previous implementation of the visibility toggle for the skin model was causing a color discrepancy issue when the model was
toggled off then had the lock body voice command called, then toggled back on. When toggling the model back on, the model's
color would reflect its color from before the toggle off command was issued, and the models color would not reflect the color change that
resulted from "lock body" command that was issued while the model was toggled off.

To resolve this issue:
- Modified the 'OnTurnModelOFF' function to correctly preserve the visible material by assigning 'visible_mat' to 'spine_mat' asset when toggling the model off. This ensures that the spine retains its correct color when toggled back on.